refactor: move validateNonAsyncFunction to utils and consolidate tests
