+++ packages/plugin-rsc/src/transforms/wrap-export.test.ts
@@ -304,58 +304,4 @@ export default Page;
       "
     `)
   })
-
-  test('reject non async function', async () => {
-    // next.js's validataion isn't entirely consisten.
-    // for now we aim to make it at least as forgiving as next.js.
-
-    const accepted = [
-      `export async function f() {}`,
-      `export default async function f() {}`,
-      `export const fn = async function fn() {}`,
-      `export const fn = async () => {}`,
-      `export const fn = async () => {}, fn2 = x`,
-      `export const fn = x`,
-      `export const fn = x({ x: y })`,
-      `export const fn = x(async () => {})`,
-      `export default x`,
-      `const y = x; export { y }`,
-      `export const fn = x(() => {})`, // rejected by next.js
-      `export const testAction = actionClient.action(async () => { return { message: "Hello, world!" }; });`,
-    ]
-
-    const rejected = [
-      `export function f() {}`,
-      `export default function f() {}`,
-      `export const fn = function fn() {}`,
-      `export const fn = () => {}`,
-      `export const fn = x, fn2 = () => {}`,
-      `export class Cls {}`,
-      `export const Cls = class {}`,
-      `export const Cls = class Foo {}`,
-    ]
-
-    async function toActual(input: string) {
-      try {
-        await testTransform(input, {
-          rejectNonAsyncFunction: true,
-        })
-        return [input, true]
-      } catch (e) {
-        return [input, e instanceof Error ? e.message : e]
-      }
-    }
-
-    const actual = [
-      ...(await Promise.all(accepted.map((e) => toActual(e)))),
-      ...(await Promise.all(rejected.map((e) => toActual(e)))),
-    ]
-
-    const expected = [
-      ...accepted.map((e) => [e, true]),
-      ...rejected.map((e) => [e, 'unsupported non async function']),
-    ]
-
-    expect(actual).toEqual(expected)
-  })
 })
